mate_fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain==0.1.16
openai==1.30.1
gtts==2.5.1
playsound==1.3.0